feat(admin/libs&&views): add vue mixin, improve admin UI

+++ admin/src/views/Banner/BannerList.vue
@@ -43,7 +43,7 @@ export default {
       data: [],
       fields: {
         _id: { label: "ID", width: "240" },
-        "bannar.name": { label: "名称", width: "" }
+        "banner.name": { label: "名称", width: "" }
       },
       pagination: {
         total: 0,
@@ -69,8 +69,29 @@ export default {
     },
 
     async remove(id) {
-      await this.$http.delete(`/rest/${this.resource.name}/${id}`);
-      this.fetch();
+      this.$confirm(
+        `此操作将永久删除${this.resource.title}, 是否继续?`,
+        "提示",
+        {
+          confirmButtonText: "确定",
+          cancelButtonText: "取消",
+          type: "warning"
+        }
+      )
+        .then(async () => {
+          await this.$http.delete(`/rest/${this.resource.name}/${id}`);
+          this.fetch();
+          this.$message({
+            type: "success",
+            message: "删除成功!"
+          });
+        })
+        .catch(() => {
+          this.$message({
+            type: "info",
+            message: "已取消删除"
+          });
+        });
     },
 
     changePageSize(pageSize) {
