+++ admin/src/views/User/UserList.vue
@@ -75,8 +75,29 @@ export default {
     },
 
     async remove(id) {
-      await this.$http.delete(`/rest/${this.resource.name}/${id}`);
-      this.fetch();
+      this.$confirm(
+        `此操作将永久删除${this.resource.title}, 是否继续?`,
+        "提示",
+        {
+          confirmButtonText: "确定",
+          cancelButtonText: "取消",
+          type: "warning"
+        }
+      )
+        .then(async () => {
+          await this.$http.delete(`/rest/${this.resource.name}/${id}`);
+          this.fetch();
+          this.$message({
+            type: "success",
+            message: "删除成功!"
+          });
+        })
+        .catch(() => {
+          this.$message({
+            type: "info",
+            message: "已取消删除"
+          });
+        });
     },
 
     changePageSize(pageSize) {
